Create QKC_R0.5
Given that QKCs operate fundamentally different from those that govern traditional kinematic couplings that contradict traditional couplings, the overconstrained problem can be solved from this point and onwards, as proposed by Martin under a set number of circumstances. 

The model has limited convergence, but once we ensure that it satisfies low frictionless coefficients, we can integrate the stress profile to obtain the force between the cone and the grooves. After which, the Culpepper transformation is applied to resultant force relative to the joint's frame of reference, allowing us to model the relationship between the surface stresses and force per unit length. 

The alternative, explored using FEA analysis, can be used instead in a more accessible manner.